sortlist-analyzer/requirements.txt
pandas
numpy
scikit-learn
networkx
duckdb
xlsxwriter
pyyaml
pytest 
openpyxl
matplotlib
seaborn